requirements.txt
asgiref==3.3.4
Django==3.2.3
django-environ==0.4.5
django-split-settings==1.0.1
paho-mqtt==1.5.1
pytz==2021.1
sqlparse==0.4.1
